<refactor> 거래 상세페이지-채팅페이지 navigate & props으로 nickname 정보 전달
