+++ src/Pages/Chat/ChatList/index.tsx
@@ -3,15 +3,15 @@ import { useState } from 'react';
 import styled from 'styled-components';
 import CommonHeader from '@components/common/CommonHeader/CommonHeader';
 import ChatListHeader from '@components/Chat/ChatList/ChatListHeader';
-import ChatLists from '@components/Chat/ChatList/ChatLists';
+import ChatLists from '@components/Chat/ChatList/ChatLists/ChatLists';
 import Character_circle from '@Assets/Character_Icons/Character_circle.svg';
 import useModal from '@hooks/useModal';
 import ConfirmModal from '@components/common/ConfirmModal';
 
   const tempData = [
-    { nickname: '동그란 당근', time: '오전 10:01', text: '시간 언제가 괜찮으신가요? 전 이번주...', userImage: Character_circle },
-    { nickname: '세모난 수박', time: '오후 4:29', text: '안녕하세요~', userImage: Character_circle },
-    { nickname: '네모난 감자', time: '오전 7:31', text: '아직 판매 하시나요?', userImage: Character_circle },
+    { id: 'round-carrot', nickname: '동그란 당근', time: '오전 10:01', text: '시간 언제가 괜찮으신가요? 전 이번주...', userImage: Character_circle },
+    { id: 'triangular-watermelon', nickname: '세모난 수박', time: '오후 4:29', text: '안녕하세요~', userImage: Character_circle },
+    { id: 'square-potato', nickname: '네모난 감자', time: '오전 7:31', text: '아직 판매 하시나요?', userImage: Character_circle },
   ];
 
 const ChatList = () => {
